+++ MatrixSDK/Utils/MXHTTPClient.m
@@ -273,7 +273,8 @@ NSString* const kMXHTTPClientMatrixErrorNotificationErrorKey = @"kMXHTTPClientMa
                     {
                         // Extract values from the home server JSON response
                         MXError *mxError = [self mxErrorFromJSON:JSONResponse];
-                        
+                        mxError.httpResponse = response;
+
                         // Send a notification
                         [[NSNotificationCenter defaultCenter] postNotificationName:kMXHTTPClientMatrixErrorNotification
                                                                             object:self