+++ CHANGES.rst
@@ -1,3 +1,14 @@
+Changes in Matrix iOS SDK in 0.13.1 (2019-07-)
+===============================================
+
+Improvements:
+ * MXError: Expose httpResponse.
+ * Soft logout: Handle new CS API error code (vector-im/riot-ios/issues/2584).
+
+Bug Fix:
+
+API break:
+
 Changes in Matrix iOS SDK in 0.13.0 (2019-07-16)
 ===============================================